requirements.txt
altgraph==0.17.4
asgiref==3.8.1
distlib==0.3.8
Django==5.0.4
filelock==3.14.0
holidays==0.47
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
pywin32-ctypes==0.2.2
setuptools==69.0.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.2
